feat: add span method to SeriesCache and its subclasses
- Implemented `span` method in `SeriesCache`, `DeltaTimeSeriesCache`, and `KVSTimeSeriesCache` to return the overall time span of the cached data.
- Updated `GroupByTimeEvent` to include parameters for `boundary_gap_ms` and `channel_time_unit`, enhancing the slicing functionality.
- Modified `determine_events` method in `GroupByTimeEvent` to utilize the new expression for slicing based on channel data span.
- Adjusted tests to validate the new functionality, ensuring correct slice boundaries and unique event instance IDs.
- Changed `event_instance_id` field type in `EVENT_INSTANCE_FACT_SCHEMA` from `IntegerType` to `LongType` for better compatibility with timestamp data.

+++ src/impulse_reporting/persist/fact_schema.py
@@ -42,7 +42,7 @@ HISTOGRAM2D_FACT_SCHEMA = StructType(
 EVENT_INSTANCE_FACT_SCHEMA = StructType(
     [
         StructField("container_id", IntegerType(), False),
-        StructField("event_instance_id", IntegerType(), False),
+        StructField("event_instance_id", LongType(), False),
         StructField("event_id", IntegerType(), False),
         StructField("start_ts", LongType(), False),
         StructField("end_ts", LongType(), False),
